pyproject.toml
[project]
name = "langchain-b12"
version = "0.1.2"
description = "A reusable collection of tools and implementations for Langchain"
readme = "README.md"
authors = [
    { name = "Vincent Min", email = "vincent.min@b12-consulting.com" }
]
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.60",
]

[dependency-groups]
google = [
    "google-genai>=1.16.1",
]
dev = [
    "pytest>=8.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = ["__pycache__", "**/scripts/*", "**/notebooks/*"]
show-fixes = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
select = [
    "F",   # Pyflakes
    "W",   # pycodestyle
    "E",   # pycodestyle
    "I",   # isort
    "UP",  # pyupgrade
    "PL",  # Pylint
    "T201", # Print Statement
]
ignore = [
    # Ignore complexity
    "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
    "EM101",
    "PLR2004",
    # Ignore f literals in exceptions
    "EM102",
]

# Add Pyright configuration
[tool.pyright]
reportUnknownParameterType = false
reportUnknownMemberType = false
reportUnknownArgumentType = false

